Split runtime construction boundaries
